LPD-16348 Move isItemHidden to utils
